add separate rule for changing underlying type
Fixes #23902.
