+++ src/Compatibility/Microsoft.DotNet.ApiCompatibility/Resources.resx
@@ -208,6 +208,9 @@
     <value>Type '{0}' is sealed because it has no visible constructor on {1} but it does on {2}</value>
   </data>
   <data name="EnumValuesMustMatch" xml:space="preserve">
-    <value>Field '{1}' of enum '{0}' changed its value from '({2}){3}' to '({4}){5}'.</value>
+    <value>Value of field '{1}' in enum '{0}' changed from '{2}' to '{3}'.</value>
+  </data>
+  <data name="EnumTypesMustMatch" xml:space="preserve">
+    <value>Underlying type of enum '{0}' changed from '{1}' to '{2}'.</value>
   </data>
 </root>
