+++ src/Compatibility/Microsoft.DotNet.ApiCompatibility/xlf/Resources.de.xlf
@@ -67,9 +67,14 @@
         <target state="translated">Das Element darf nicht NULL sein. NULL gefunden bei Index: {0}.</target>
         <note />
       </trans-unit>
+      <trans-unit id="EnumTypesMustMatch">
+        <source>Underlying type of enum '{0}' changed from '{1}' to '{2}'.</source>
+        <target state="new">Underlying type of enum '{0}' changed from '{1}' to '{2}'.</target>
+        <note />
+      </trans-unit>
       <trans-unit id="EnumValuesMustMatch">
-        <source>Field '{1}' of enum '{0}' changed its value from '({2}){3}' to '({4}){5}'.</source>
-        <target state="new">Field '{1}' of enum '{0}' changed its value from '({2}){3}' to '({4}){5}'.</target>
+        <source>Value of field '{1}' in enum '{0}' changed from '{2}' to '{3}'.</source>
+        <target state="new">Value of field '{1}' in enum '{0}' changed from '{2}' to '{3}'.</target>
         <note />
       </trans-unit>
       <trans-unit id="FileDoesNotExist">
